requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-autofixture==0.12.1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.3
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.16.2
pandas==0.24.2
psycopg2==2.8.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.2
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
